FeeWallet trait
