+++ stacks-js-rs/Cargo.toml
@@ -5,5 +5,6 @@ edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 [dependencies]
+stacks-coordinator = { path = "../stacks-coordinator" }
 serde_json = { workspace = true }
 blockstack-core = { workspace = true }
